fix(dashboard): autosave for the code created steps - custom controls
